cleanup: refactor to simplify, reduce imports
