+++ apps/content-watcher/src/controllers/health.controller.ts
@@ -4,6 +4,7 @@ import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { HealthCheckService } from '#health-check/health-check.service';
 import { HealthResponseDto } from '#types/dtos/common/health.response.dto';
 import { ContentWatcherQueues as QueueConstants } from '#types/constants/queue.constants';
+import { IContentWatcherApiConfig } from '#content-watcher/api.config';
 
 @Controller()
 @ApiTags('health')
@@ -21,7 +22,7 @@ export class HealthController {
   @ApiOkResponse({ description: 'Service is healthy' })
   async healthz(): Promise<HealthResponseDto> {
     const [configResult, redisResult, blockchainResult] = await Promise.allSettled([
-      this.healthCheckService.getServiceConfig('content-watcher-api'),
+      this.healthCheckService.getServiceConfig<IContentWatcherApiConfig>('content-watcher-api'),
       this.healthCheckService.getRedisStatus(QueueConstants.QUEUE_NAMES),
       this.healthCheckService.getBlockchainStatus(),
     ]);
