+++ libs/health-check/src/health-check.service.ts
@@ -4,14 +4,26 @@ import { BlockchainRpcQueryService } from '#blockchain/blockchain-rpc-query.serv
 import { InjectRedis } from '@songkeys/nestjs-redis';
 import Redis from 'ioredis';
 import { RedisStatusDto, QueueStatusDto, BlockchainStatusDto, LatestBlockHeader } from '#types/dtos/common';
-import { ConfigTypeName, ConfigType, QueueName } from './types';
-import { QUEUE_NAME_TO_REDIS_KEY_MAP } from './constants';
 
 import { plainToInstance } from 'class-transformer';
 import fs from 'fs';
 
 import { Logger, pino } from 'pino';
 import { getBasicPinoOptions } from '#logger-lib';
+
+import {
+  QUEUE_PREFIX,
+  AccountQueues,
+  ContentPublishingQueues,
+  ContentWatcherQueues,
+  GraphQueues,
+} from '#types/constants/queue.constants';
+
+type QueueName =
+  | AccountQueues.QueueName
+  | ContentPublishingQueues.QueueName
+  | ContentWatcherQueues.QueueName
+  | GraphQueues.QueueName;
 
 interface RedisWithCustomCommands extends Redis {
   queueStatus(prefixes: string[]): Promise<string>;
@@ -40,8 +52,8 @@ export class HealthCheckService {
     });
   }
 
-  public async getRedisStatus(queues: Array<QueueName>): Promise<RedisStatusDto> {
-    const queuePrefixes = queues.map((queueName) => this.QueueNameToKey(queueName)).filter((key) => !!key);
+  public async getRedisStatus(queueNames: Array<QueueName>): Promise<RedisStatusDto> {
+    const queuePrefixes = queueNames.map((queueName) => `${QUEUE_PREFIX}:${queueName}`).filter(Boolean);
     if (queuePrefixes.length === 0) {
       this.logger.warn('No valid queue prefixes found for status check');
     }
@@ -76,14 +88,6 @@ export class HealthCheckService {
     });
   }
 
-  private QueueNameToKey(queueName: QueueName): string {
-    const redisKey = QUEUE_NAME_TO_REDIS_KEY_MAP[queueName];
-    if (!redisKey) {
-      this.logger.warn(`Queue not found: ${queueName}`);
-    }
-    return redisKey;
-  }
-
   public async getBlockchainStatus(): Promise<BlockchainStatusDto> {
     return plainToInstance(BlockchainStatusDto, {
       frequencyApiWsUrl: this.configService.get<string>('FREQUENCY_API_WS_URL'),
@@ -113,7 +117,7 @@ export class HealthCheckService {
     }
   }
 
-  public getServiceConfig<T extends ConfigTypeName>(serviceName: T): ConfigType<T> {
-    return this.configService.get<ConfigType<T>>(serviceName);
+  public getServiceConfig<T = unknown>(serviceName: string): T {
+    return this.configService.get<T>(serviceName);
   }
 }
