+++ libs/types/src/dtos/account/index.ts
@@ -12,4 +12,3 @@ export * from './transaction.response.dto';
 export * from './wallet.login.config.response.dto';
 export * from './wallet.login.request.dto';
 export * from './wallet.login.response.dto';
-export * from './config.dto';
