+++ libs/types/src/dtos/content-publishing/index.ts
@@ -2,4 +2,3 @@ export * from './activity.dto';
 export * from './announcement.dto';
 export * from './common.dto';
 export * from './validation';
-export * from './config.dto';
