--- libs/types/src/interfaces/content-watcher/api-config.interface.ts
@@ -1,5 +1,0 @@
-export interface IContentWatcherApiConfig {
-  apiBodyJsonLimit: string;
-  apiPort: number;
-  apiTimeoutMs: number;
-}
