+++ libs/types/src/interfaces/content-watcher/index.ts
@@ -2,4 +2,3 @@ export * from './announcement-subscription.interface';
 export * from './ipfs.job.interface';
 export * from './message_response_with_schema_id';
 export * from './scan-reset.interface';
-export * from './api-config.interface';
