--- libs/types/src/interfaces/graph/api-config.interface.ts
@@ -1,5 +1,0 @@
-export interface IGraphApiConfig {
-  apiBodyJsonLimit: string;
-  apiPort: number;
-  apiTimeoutMs: number;
-}
