+++ libs/types/src/interfaces/graph/index.ts
@@ -1,3 +1,2 @@
 export * from './capacity-limit.interface';
 export * from './provider.graph.update-job.interface';
-export * from './api-config.interface';
